pyproject.toml
[tool.poetry]
name = "metealologia-api"
version = "0.1.0"
description = "API service for the Metealologia project"
authors = ["Asapros <asapros@wp.pl>"]
readme = "README.md"
packages = [{include = "metealologia_api"}]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.23"

[tool.poetry.scripts]
start = "metealologia_api.main:start"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"